\name{bmi-package}
\alias{bmi-package}
\alias{bmi}
\docType{package}
\title{
\packageTitle{bmi}
}
\description{
\packageDescription{bmi}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{bmi}

Maintainer: \packageMaintainer{bmi}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = bmi)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{bmi}}
%\section{Documentation Index}{\packageIndices{bmi}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
